chore(vm): refactor: move disks code out of vm.go
