+++ proxmoxtf/resource/validators/network.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package validator
+package validators
 
 import (
 	"fmt"
